+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 * The process for updating Tor on Android is now also mostly automated and part of our release checklist
 * Stops duplicating Tor binaries on MacOS (we were just duplicating the same Universal Binary for both x64 and arm64, which did not make sense.)
 
+### Fixes
+
+* Changing the text on the loading screen when the user creates a community. [#2248] (https://github.com/TryQuiet/quiet/issues/2248)
+
 ## [5.1.0]
 
 ### Chores