+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Features
 
 * Adds a hook for QSS to inject entries into OrbitDB [#2807](https://github.com/TryQuiet/quiet/issues/2807)
+* Adds an event for when a local put occurs [#2802](https://github.com/TryQuiet/quiet/issues/2802)
 
 ### Fixes
 
 * Fixes race condition with initial user profile entry on joining [#2887](https://github.com/TryQuiet/quiet/issues/2887)
+* Changing the text on the loading screen when the user creates a community. [#2248] (https://github.com/TryQuiet/quiet/issues/2248)
 
 ### Chores
 
@@ -18,9 +20,6 @@
 * The process for updating Tor on Android is now also mostly automated and part of our release checklist
 * Stops duplicating Tor binaries on MacOS (we were just duplicating the same Universal Binary for both x64 and arm64, which did not make sense.)
 
-### Fixes
-
-* Changing the text on the loading screen when the user creates a community. [#2248] (https://github.com/TryQuiet/quiet/issues/2248)
 
 ## [5.1.0]